refactor(storage): delegate to domain repositories
